+++ front-end/src/__fixtures__/gpuServersResponses.ts
